+++ public/js/app.js
@@ -5,15 +5,15 @@
 // ═══════════════════════════════════════════════════════════
 
 import SocketMethods   from './modules/app-socket.js?v=4.6.2';
-import UIBindMethods   from './modules/app-ui.js?v=4.6.2';
+import UIBindMethods   from './modules/app-ui.js?v=4.6.2b';
 import MediaMethods    from './modules/app-media.js?v=4.6.2';
 import ContextMethods  from './modules/app-context.js?v=4.6.2';
 import ChannelMethods  from './modules/app-channels.js?v=4.6.2';
 import MessageMethods  from './modules/app-messages.js?v=4.6.2';
 import UserMethods     from './modules/app-users.js?v=4.6.1';
-import VoiceMethods    from './modules/app-voice.js?v=4.5.1';
+import VoiceMethods    from './modules/app-voice.js?v=4.6.2';
 import UtilityMethods  from './modules/app-utilities.js?v=4.6.2';
-import AdminMethods    from './modules/app-admin.js?v=4.6.2';
+import AdminMethods    from './modules/app-admin.js?v=4.6.2b';
 import PlatformMethods from './modules/app-platform.js?v=4.5.1';
 import SearchMethods   from './modules/app-search.js?v=3.49.0';
 import FerryMethods    from './modules/app-ferry.js?v=3.51.4';